HDF5: CMake 3.6 fix Components
Due to a fix in the default settings of HDF5
components, C++ projects compiling against the C-only
version of HDF5 might fail with

```
CMake Error: The following variables are used in this project, but they are set to NOTFOUND.
Please set them or make sure they are set and tested correctly in the CMake files:
HDF5_hdf5_cpp_LIBRARY
```

This fixes it by setting exactly the component
we need from HDF5.

Versions before CMake 3.5 are not affected and choose
the C bindings only (at least since 2.8.11).